Updated packages and fixed errors to ensure project runs smoothly
I was trying to run this project to update with some new information but I found a lot of errors and packages not updated, so I decided to update the packages and fix the erros to be able to run the project again.

I made some tests and the project is running with all the updates without any error.

I'll do another PR to add the information that I was talking about before.